add mathod to analyze json
