+++ Makefile
@@ -1,13 +1,14 @@
 CC := g++
 
-json:utils.o read.o jsonobj.o jsonarray.o
+json:utils.o read.o jsonobj.o jsonarray.o dataadapter.o
 	$(CC) -o $@ $^
 
 utils.o:utils.h
-read.o: jsonobj.h jsonarray.h
-jsonobj.o:jsonobj.h dataadapter.h utils.h
-jsonarray.o:jsonarray.h dataadapter.h utils.h
+jsonobj.o:json.h utils.h
+jsonarray.o:json.h utils.h
+dataadapter.o:json.h
+read.o: json.h
 
 .PHONY : clean
 clean:
-	rm json utils.o read.o jsonobj.o jsonarray.o -r
+	rm json utils.o read.o jsonobj.o jsonarray.o dataadapter.o -f
